Prevent dynamic data from being output on cached pages
Use the gtm_server_side_is_page_cached filter (provided by the user) to determine whether it’s safe to output dynamic or sensitive data.

As an additional safeguard, we insert an HTML comment '<!-- DO NOT CACHE -->' whenever such data is output. This comment can be used to scan for potential data leaks in cached content.

At a minimum, users can include a check for this comment in their monitoring systems and trigger an alert or failure if it’s detected on a page that should be cached, such as the homepage.